tools/ebpf-check: Process multiple .o files, all TEXT sections
Bring a few changes to the list of arguments for the Python script:

- The argument "--sections" now takes a space-separated list (instead of
  a comma-separated list) of sections, we leave it to argparse to get
  the list.

- Change the default behaviour: Process all TEXT sections (as showned by
  llvm-objdump, and in which eBPF instructions are usually stored) if no
  section name is provided on the command line, instead of .text.

- Allow passing several object files to process.

Remove the shell script to run ebpf-check.py on all files of a
directory, given that this is now trivial to do with a shell wildcard.
We no longer use llvm-objdump to retrieve all section names; remove it
from the list of prerequisites in the README file.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>