tools/ebpf-check: Parse description when generating list of instructions
In the future, we will likely stop to parse the Appendix table from the
RST document to dynamically generate the list of instructions, and will
instead have a reference JSON committed into the repository.

In preparation for that, let's add the description and category of each
instruction to the generated JSON, to help human read and fix it in the
future if necessary.

Head of the generated file with the current RST:

    [
      {
        "opc": 0,
        "src": 0,
        "imm": "any",
        "dsc": "(additional immediate value)",
        "cat": "64-bit immediate instructions"
      },
      {
        "opc": 4,
        "src": 0,
        "imm": "any",
        "dsc": "dst = (uint32_t)(dst + imm)",
        "cat": "Arithmetic instructions"
      },
      ...
    ]

Signed-off-by: Quentin Monnet <quentin@isovalent.com>