tools/ebpf-check: Add check on 2nd-half imm values for lddw extensions
The extensions based on the 128-bit long instruction initially used to
load double words may have some constraints on the value of the
immediate in their second half. Because we process all other
instructions as 8-byte long, we don't account for that in our list of
valid instructions: we only have one entry for second halves, starting
with opcode 0x00, and taking any value for the immediate.

Add a check on the immediate in the second half-instruction for the
relevant instructions.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>